requirement.txt
pytorch==1.12.1
rdkit==2023.3.1
matplotlib
seaborn